+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.4] - 2026-02-07
+
+### Added
+
+- Simplified parameters
+- Introduced config file
+- Added feature to crop output to a square centered on the map
+
 ## [1.7.3] - 2025-12-13
 
 ### Fixed